Create CVE-2021-40444.md
# Microsoft MSHTML Remote Code Execution Vulnerability (CVE-2021-40444)

## Author
McAfee

## Description
This expert rule detects abuse of MSHTML Remote Code Execution Vulnerability through Microsoft Word application. This is a stop-gap solution and this rule is expected to false if Customer environment is dependent on Winword.exe using mshtml.dll for operations.

## Rule Class 
Processes

## Rule TCL
```tcl
The original rule: 
Rule {​​​​​​​
    Process {​​​​​​​
        Include OBJECT_NAME {​​​​​​​ -v "winword.exe" }​​​​​​​
        Include DLL_LOADED -name "ieframe" {​​​​​​​ -v 0x1 }​​​​​​​
    }​​​​​​​
    Target {​​​​​​​
        Match SECTION {​​​​​​​
            Include OBJECT_NAME {​​​​​​​ -v "mshtml.dll" }​​​​​​​
        }​​​​​​​
    }​​​​​​​
}​​​​​​​   
```

## Trigger
NA

## Tested Platforms
OS: Windows 11 x64 and Win 10 21H1 x86
ENS: 10.7.0 Jul'21 update

## Notes
Customers are advised to fine-tune the rule in their environment or disable the signature if there are false positives.
We have observed False Positives while signing in to Office (winword) application. We are working on improving this Expert rule.